Do not pass proxy options to Net::HTTP connection if not specified
Ruby 2.0 has changed the Net::HTTP.new method so that if you don't specify
a proxy address it defaults to using the proxy specified by the http_proxy
environment variable (if set). Passing nil for the proxy address now means
that a proxy is never used even if the http_proxy environment variable is
set.

However HTTParty was always passing the proxy option values even if they
were nil, which meant that the defaulting to http_proxy environment variable
didn't work.

With this change the proxy values are only passed if http_proxyaddr is non-nil.